VQ3D/requirements.txt
imageio==2.9.0
numpy==1.19.2
open3d==0.13.0
opencv-python==4.5.1.48
pandas
pillow
torch==1.8.1
torchvision==0.9.1
timm==0.4.5
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.2
tqdm==4.60.0
timm==0.4.5
